Include a copy of platforms.svg
Made this at the request of SteamDB.info, they were searching for a
better looking 16×16px image of the Linux platform icon, and I offered
an adjusted decapitated variant of tuxfonty.svg
This was done at the time of 0882fd684be49aecc7d5d0d22ddc51dbfe7c1eed
but I didn’t include it ‐ wasn’t sure if I wanted it in the repo, but
eh. Why not.
Will sync gh-pages branch with a few examples of this whenever.